# packages/contracts/openapi/social-platform.v1.yaml
openapi: 3.1.0
info:
  title: Swypik Social Platform API
  version: 1.0.0
  description: Modular monorepo platform API contract for feed, events, video, social, cart, checkout, payments, and moderation.
servers:
  - url: http://localhost:8080
    description: Local platform-api
...
paths:
...
  /v1/feed:
    get:
      tags: [feed]
      operationId: listFeed
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: category_id
          in: query
          schema:
            type: string
        - name: creator_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ranked feed page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
  /v1/events/batch:
    post:
      tags: [events]
      operationId: recordEventBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBatchRequest"
      responses:
        "202":
          description: Events accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/videos/uploads/init:
    post:
      tags: [videos]
      operationId: initializeVideoUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInitRequest"
      responses:
        "201":
          description: Upload URL initialized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadInitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/videos/uploads/complete:
    post:
      tags: [videos]
      operationId: completeVideoUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadCompleteRequest"
      responses:
        "200":
          description: Upload completed and video created.
          content:
            application/json:
              schema:
                type: object
                required: [video]
                properties:
                  video:
                    $ref: "#/components/schemas/Video"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/videos/uploads/{id}/status:
    get:
      tags: [videos]
      operationId: getVideoUploadStatus
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Upload status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/videos/{id}/publish:
    post:
      tags: [videos]
      operationId: publishVideo
      parameters:
        - $ref: "#/components/parameters/PathID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishVideoRequest"
      responses:
        "200":
          description: Video published.
          content:
            application/json:
              schema:
                type: object
                required: [video]
                properties:
                  video:
                    $ref: "#/components/schemas/Video"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/social/follow:
    post:
      tags: [social]
      operationId: followUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowRequest"
      responses:
        "200":
          description: Follow state updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/social/unfollow:
    post:
      tags: [social]
      operationId: unfollowUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowRequest"
      responses:
        "200":
          description: Follow state updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/videos/{id}/comments:
    get:
      tags: [videos]
      operationId: listVideoComments
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Video comments.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
    post:
      tags: [videos]
      operationId: addVideoComment
      parameters:
        - $ref: "#/components/parameters/PathID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/cart/items:
    post:
      tags: [checkout]
      operationId: addCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Cart item added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/checkout:
    post:
      tags: [checkout]
      operationId: createCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Checkout session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/payments/webhooks/stripe:
    post:
      tags: [checkout]
      operationId: recordStripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Stripe event received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeWebhookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource is in a conflicting state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    FeedResponse:
      type: object
      required: [items, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedItem"
        limit:
          type: integer
        offset:
          type: integer
    FeedItem:
      type: object
      required: [id, has_video, rank_score]
      properties:
        id:
          type: string
        product_id:
          type: string
        creator_id:
          type: string
        category_id:
          type: string
        title:
          type: string
        description:
          type: string
        has_video:
          type: boolean
        video_url:
          type: string
          format: uri
        poster_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        price_ron:
          type: integer
        old_price_ron:
          type: integer
        rating:
          type: number
        orders_count:
          type: integer
        likes_count:
          type: integer
        views_count:
          type: integer
        comments_count:
          type: integer
        published_at:
          type: string
          format: date-time
        rank_score:
          type: number
    EventBatchRequest:
      type: object
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/Event"
    Event:
      type: object
      required: [type, subject_type, subject_id]
      properties:
        id:
          type: string
        type:
          type: string
          minLength: 1
        event_type:
          type: string
          deprecated: true
        actor_id:
          type: string
        subject_type:
          type: string
          minLength: 1
        subject_id:
          type: string
          minLength: 1
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    EventBatchResponse:
      type: object
      required: [batch_id, accepted, stored]
      properties:
        batch_id:
          type: string
        accepted:
          type: integer
        stored:
          type: boolean
    UploadInitRequest:
      type: object
      required: [creator_id, filename, content_type, size_bytes]
      properties:
        creator_id:
          type: string
          minLength: 1
        filename:
          type: string
          minLength: 1
        content_type:
          type: string
          pattern: "^video/"
        size_bytes:
          type: integer
          minimum: 1
          maximum: 1073741824
        product_id:
          type: string
        checksum_sha:
          type: string
        original_name:
          type: string
    UploadInitResponse:
      type: object
      required: [upload_id, upload_url, method, headers, expires_at, status]
      properties:
        upload_id:
          type: string
        upload_url:
          type: string
          format: uri
        method:
          type: string
          const: PUT
        headers:
          type: object
          additionalProperties:
            type: string
        expires_at:
          type: string
          format: date-time
        status:
          type: string
          const: pending
    UploadCompleteRequest:
      type: object
      required: [upload_id, creator_id, video_url]
      properties:
        upload_id:
          type: string
        creator_id:
          type: string
        video_url:
          type: string
          format: uri
        poster_url:
          type: string
          format: uri
        duration_ms:
          type: integer
          minimum: 0
        storage_path:
          type: string
    UploadStatus:
      type: object
      required: [id, creator_id, filename, content_type, size_bytes, status, expires_at, created_at]
      properties:
        id:
          type: string
        creator_id:
          type: string
        product_id:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
          enum: [pending, completed]
        video_id:
          type: string
        upload_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Video:
      type: object
      required: [video_id, id, creator_id, video_url, status, created_at]
      properties:
        video_id:
          type: string
        id:
          type: string
        upload_id:
          type: string
        creator_id:
          type: string
        product_id:
          type: string
        video_url:
          type: string
          format: uri
        poster_url:
          type: string
          format: uri
        duration_ms:
          type: integer
        status:
          type: string
          enum: [ready, published]
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    PublishVideoRequest:
      type: object
      required: [creator_id]
      properties:
        creator_id:
          type: string
    FollowRequest:
      type: object
      required: [follower_id, followee_id]
      properties:
        follower_id:
          type: string
        followee_id:
          type: string
    FollowResponse:
      type: object
      required: [follower_id, followee_id, following, updated_at]
      properties:
        follower_id:
          type: string
        followee_id:
          type: string
        following:
          type: boolean
        updated_at:
          type: string
          format: date-time
    AddCommentRequest:
      type: object
      required: [author_id, body]
      properties:
        author_id:
          type: string
        body:
          type: string
          minLength: 1
          maxLength: 2000
    Comment:
      type: object
      required: [id, video_id, author_id, body, created_at]
      properties:
        id:
          type: string
        video_id:
          type: string
        author_id:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
    AddCartItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        cart_id:
          type: string
        user_id:
          type: string
        product_id:
          type: string
        sku_id:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 20
        unit_amount_ron:
          type: integer
          minimum: 0
    CartItem:
      allOf:
        - $ref: "#/components/schemas/AddCartItemRequest"
        - type: object
          required: [id, cart_id, product_id, quantity, unit_amount_ron, created_at]
          properties:
            id:
              type: string
            created_at:
              type: string
              format: date-time
    CheckoutRequest:
      type: object
      properties:
        cart_id:
          type: string
        user_id:
          type: string
        currency:
          type: string
          default: RON
        items:
          type: array
          items:
            $ref: "#/components/schemas/AddCartItemRequest"
        customer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
        metadata:
          type: object
          additionalProperties: true
    CheckoutResponse:
      type: object
      required: [checkout_id, status, payment_url, currency, total_amount_ron, items]
      properties:
        checkout_id:
          type: string
        status:
          type: string
          enum: [requires_payment]
        payment_url:
          type: string
          format: uri
        currency:
          type: string
        total_amount_ron:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
    StripeWebhookResponse:
      type: object
      required: [received]
      properties:
        received:
          type: boolean
        event_id:
          type: string
        event_type:
          type: string
...